remove invalid subtraction from release discard handler
The subtraction was necessary when initially planned as the parsing of the release ids from the directories was happening after the check. Now that the ids are parsed while filtering the initial read result, there is no need anymore to remove one directory as the ".easydep_base_repo" directory name cannot be parsed as an u64 and will therefore get filtered out prior to the check.